J: Rate of electron transport (umol/m^2/s)

Calculate the rate of electron transport as a function of photosynthetic photon flux density (PPFD).

Usage

J(pars, unitless = FALSE)

Arguments

pars

Concatenated parameters (leaf_par, enviro_par, and constants)

unitless

Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

Value in \(\mu\)mol/ (m^2 s) of class units

Details

\(J\) as a function of PPFD is the solution to the quadratic expression:

$$0 = \theta_J J ^ 2 - J (J_\mathrm{max} + \phi_J PPFD) + J_\mathrm{max} \phi_J PPFD$$

Symbol	R	Description	Units	Default

\(J_\mathrm{max}\)	J_max	potential electron transport (T_leaf)	\(\mu\)mol CO2 / (m\(^2\) s)	calculated

\(\phi_J\)	phi_J	initial slope of the response of J to PPFD	none	0.331

PPFD	PPFD	photosynthetic photon flux density	\(\mu\)mol quanta / (m^2 s)	1500

\(\theta_J\)	theta_J	curvature factor for light-response curve	none	0.825

Examples

library(magrittr)
library(photosynthesis)

bake_par = make_bakepar()
constants = make_constants(use_tealeaves = FALSE)
enviro_par = make_enviropar(use_tealeaves = FALSE)
leaf_par = make_leafpar(use_tealeaves = FALSE)
enviro_par$T_air = leaf_par$T_leaf
leaf_par %<>% bake(enviro_par, bake_par, constants)

pars = c(leaf_par, enviro_par, constants)
J(pars, FALSE)
#> 181.6577 [umol/m^2/s]