Non-rectangular hyperbolic model of light responses

Please use marshall_biscoe_1980().

Usage

aq_response(k_sat, phi_J, Q_abs, theta_J)

Arguments

k_sat

Light saturated rate of process k

phi_J

Quantum efficiency of process k

Q_abs

Absorbed light intensity (umol m-2 s-1)

theta_J

Curvature of the light response

Value

aq_response is used to describe the response of a process to absorbed light intensity. Assumes that input is absorbed light. Note that if absorbed light is not used, then the meaning of phi_J becomes unclear. This function is designed to be used with fit_aq_response, however it could easily be fed into a different fitting approach (e.g. Bayesian approaches). Originally from Marshall et al. 1980.

References

Marshall B, Biscoe P. 1980. A model for C3 leaves describing the dependence of net photosynthesis on irradiance. J Ex Bot 31:29-39