Get default model

Get the name of the default model used for different plant ecophysiological data analysis methods implemented in photosynthesis. Currently only used for fit_aq_response2() and fit_r_light2().

Light response models:

marshall_biscoe_1980(): Non-rectangular hyperbolic model of light responses

photoinhibition(): Non-rectangular hyperbolic model of light responses with photoinhibition of k_sat at increasing Q_abs

Usage

get_default_model(.photo_fun)

get_all_models(method)

marshall_biscoe_1980(Q_abs, k_sat, phi_J, theta_J)

photoinhibition(Q_abs, k_sat, phi_J, theta_J, b_inh)

Arguments

.photo_fun

A character string of photosynthesis function to call. One of: aq_response, r_light.

method

A character string of the statistical method to use: 'ls' for least-squares and 'brms' for Bayesian model using brms::brm(). Default is 'ls'.

Q_abs

Absorbed light intensity (\(\mu\)mol m\(^{-2}\) s\(^{-1}\))

k_sat

Light saturated rate of process k

phi_J

Quantum efficiency of process k

theta_J

Curvature of the light response

b_inh

Inhibition parameter

Value

A character string with name of model.

Examples

get_default_model("aq_response")
#> [1] "marshall_biscoe_1980"
get_default_model("r_light")
#> [1] "walker_ort_2015"