CO2 supply and demand function (mol / m^2 s)

This function is not intended to be called by users directly.

A_supply(C_chl, pars, unitless = FALSE)

A_demand(C_chl, pars, unitless = FALSE)

Arguments

C_chl
Chloroplastic CO2 concentration in Pa of class units

pars
Concatenated parameters (leaf_par, enviro_par, and constants)

unitless
Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

Value

Value in mol / (m^2 s) of class units

Details

Supply function:
$$A = g_\mathrm{tc} (C_\mathrm{air} - C_\mathrm{chl})$$

Demand function:
$$A = (1 - \Gamma* / C_\mathrm{chl}) \mathrm{min}(W_\mathrm{carbox}, W_\mathrm{regen}, W_\mathrm{tpu}) - R_\mathrm{d}$$

Symbol	R	Description	Units	Default

\(A\)	A	photosynthetic rate	\(\mu\)mol CO2 / (m^2 s)	calculated

\(g_\mathrm{tc}\)	g_tc	total conductance to CO2	\(\mu\)mol CO2 / (m\(^2\) s Pa)	calculated

\(C_\mathrm{air}\)	C_air	atmospheric CO2 concentration	Pa	41

\(C_\mathrm{chl}\)	C_chl	chloroplastic CO2 concentration	Pa	calculated

\(R_\mathrm{d}\)	R_d	nonphotorespiratory CO2 release	\(\mu\)mol CO2 / (m\(^2\) s)	2

\(\Gamma*\)	gamma_star	chloroplastic CO2 compensation point	Pa	3.743

Examples

bake_par <- make_bakepar()
constants <- make_constants(use_tealeaves = FALSE)
enviro_par <- make_enviropar(use_tealeaves = FALSE)
leaf_par <- make_leafpar(use_tealeaves = FALSE)
leaf_par <- bake(leaf_par, bake_par, constants)
# Or bake with piping (need library(magrittr))
# leaf_par %<>% bake(bake_par, constants)
enviro_par$T_air <- leaf_par$T_leaf

pars <- c(leaf_par, enviro_par, constants)
C_chl <- set_units(35, "Pa")

A_supply(C_chl, pars)

#> 10.01259 [umol/m^2/s]

A_demand(C_chl, pars)

#> 31.41145 [umol/m^2/s]