Conductance to CO2 (umol / (m^2 s Pa))

Conductance to CO2 (umol / (m^2 s Pa))

- g_tc: total conductance to CO2

- g_uc: cuticular conductance to CO2

- g_bc: boundary layer conductance to CO2

- g_mc: mesophyll conductance to CO2

- g_sc: stomatal conductance to CO2

.get_gtc(pars, unitless)

.get_guc(pars, surface, unitless)

.get_gbc(pars, surface, unitless)

.get_gmc(pars, surface, unitless)

.get_gsc(pars, surface, unitless)

Arguments

pars
Concatenated parameters (leaf_par, enviro_par, and constants)

unitless
Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

surface
Leaf surface (lower or upper)

Details

Total conductance to CO2 is the sum of parallel conductances on the lower (\(g_\mathrm{c,lower}\)) and upper (\(g_\mathrm{c,upper}\)) leaf portions:

$$g_\mathrm{c,total} = g_\mathrm{c,lower} + g_\mathrm{c,upper}$$

Each partial conductance consists of two parallel conductances, the cuticular conductance (\(g_\mathrm{u,c}\)) and the in-series conductances through mesophyll (\(g_\mathrm{m,c}\)), stomata (\(g_\mathrm{s,c}\)), and boundary layer (\(g_\mathrm{b,c}\)). To simplify the formula, I use substitute resistance where \(r_x = 1 / g_x\). For surface \(i\):

$$g_{\mathrm{c},i} = g_{\mathrm{u},i} + (1 / (r_{\mathrm{m},i} + r_{\mathrm{s},i} + r_{\mathrm{b},i}))$$

The cuticular, stomatal, and mesophyll conductances can be the same or different for upper and lower. The partitioning factors (\(k_x\)) divide the conductance between surfaces while keeping the total conductance constant:

$$g_{x,\mathrm{lower}} = g_x (1 / (1 + k_x))$$ $$g_{x,\mathrm{upper}} = g_x (k_x / (1 + k_x))$$ $$g_x = g_{x,\mathrm{lower}} + g_{x,\mathrm{upper}}$$

How the partitioning factors work:

\(k_x\)	description

0	all conductance on lower surface/portion

0.5	2/3 conductance on lower surface

1	conductance evenly divided between surfaces/portions

2	2/3 conductance on upper surface

Inf	all conductance on upper surface/portion

The boundary layer conductances for each are calculated on the basis of mass and heat transfer (see .get_gbc).

Symbol	R	Description	Units	Default

\(g_\mathrm{mc}\)	g_mc	mesophyll conductance to CO2 (T_leaf)	\(\mu\)mol CO2 / (m\(^2\) s Pa)	calculated

\(g_\mathrm{sc}\)	g_sc	stomatal conductance to CO2	\(\mu\)mol CO2 / (m\(^2\) s Pa)	4

\(g_\mathrm{uc}\)	g_uc	cuticular conductance to CO2	\(\mu\)mol CO2 / (m\(^2\) s Pa)	0.1

\(k_\mathrm{mc}\)	k_mc	partition of \(g_\mathrm{mc}\) to lower mesophyll	none	1

\(k_\mathrm{sc}\)	k_sc	partition of \(g_\mathrm{sc}\) to lower surface	none	1

\(k_\mathrm{uc}\)	k_uc	partition of \(g_\mathrm{uc}\) to lower surface	none	1

Examples

# Total conductance to CO2

library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

library(photosynthesis)

## Hypostomatous leaf; default parameters
bake_par <- make_bakepar()
constants <- make_constants(use_tealeaves = FALSE)
enviro_par <- make_enviropar(use_tealeaves = FALSE)
leaf_par <- make_leafpar(replace = list(k_sc = set_units(0)),
                         use_tealeaves = FALSE)
leaf_par %<>% bake(bake_par, constants)
enviro_par$T_air <- leaf_par$T_leaf
pars <- c(leaf_par, enviro_par, constants)

gbw_upper <- photosynthesis:::.get_gbc(pars, "upper", FALSE)

# Lower surface ----
gbc_lower <- photosynthesis:::.get_gbc(pars, "lower", FALSE)
gmc_lower <- photosynthesis:::.get_gmc(pars, "lower", FALSE)
gsc_lower <- photosynthesis:::.get_gsc(pars, "lower", FALSE)
guc_lower <- photosynthesis:::.get_guc(pars, "lower", FALSE)

rc_lower <- 1 / gmc_lower + 1 / gsc_lower + 1 / gbc_lower
gc_lower <- 1 / rc_lower
gc_lower %<>% magrittr::add(guc_lower)

# Upper surface ----
gbc_upper <- photosynthesis:::.get_gbc(pars, "upper", FALSE)
gmc_upper <- photosynthesis:::.get_gmc(pars, "upper", FALSE)
gsc_upper <- photosynthesis:::.get_gsc(pars, "upper", FALSE)
guc_upper <- photosynthesis:::.get_guc(pars, "upper", FALSE)

rc_upper <- 1 / gmc_upper + 1 / gsc_upper + 1 / gbc_upper
gc_upper <- 1 / rc_upper
gc_upper %<>% magrittr::add(guc_upper)

## Lower and upper surface are in parallel
g_tc <- gc_lower + gc_upper

g_tc

#> 1.075716 [umol/m^2/Pa/s]