Conductance to CO2 (mol / m^2 / s)

Conductance to CO2 (mol / m^2 / s)

- g_tc: total conductance to CO2

- g_uc: cuticular conductance to CO2

- g_bc: boundary layer conductance to CO2

- g_mc: mesophyll conductance to CO2

- g_sc: stomatal conductance to CO2

.get_gtc(pars, unitless, use_legacy_version)

.get_guc(pars, surface, unitless)

.get_gbc(pars, surface, unitless, use_legacy_version)

.get_gmc(pars, surface, unitless)

.get_gsc(pars, surface, unitless)

Arguments

pars

Concatenated parameters (leaf_par, enviro_par, and constants)

unitless

Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

use_legacy_version

Logical. Should legacy model (<2.1.0) be used? See NEWS for further information. Default is FALSE.

surface

Leaf surface (lower or upper)

Details

Default conductance model

The conductance model described in this section is used by default unless additional anatomical parameters described in the next section are provided.

Total conductance to CO2 is the sum of parallel conductances on the lower (\(g_\mathrm{c,lower}\)) and upper (\(g_\mathrm{c,upper}\)) leaf portions:

$$g_\mathrm{c,total} = g_\mathrm{c,lower} + g_\mathrm{c,upper}$$

Each partial conductance consists of two parallel conductances, the cuticular conductance (\(g_\mathrm{u,c}\)) and the in-series conductances through mesophyll (\(g_\mathrm{m,c}\)), stomata (\(g_\mathrm{s,c}\)), and boundary layer (\(g_\mathrm{b,c}\)). To simplify the formula, I use substitute resistance where \(r_x = 1 / g_x\). For surface \(i\):

$$g_{\mathrm{c},i} = g_{\mathrm{u},i} + (1 / (r_{\mathrm{m},i} + r_{\mathrm{s},i} + r_{\mathrm{b},i}))$$

The cuticular, stomatal, and mesophyll conductances can be the same or different for upper and lower. The partitioning factors (\(k_x\)) divide the conductance between surfaces while keeping the total conductance constant:

$$g_{x,\mathrm{lower}} = g_x (1 / (1 + k_x))$$ $$g_{x,\mathrm{upper}} = g_x (k_x / (1 + k_x))$$ $$g_x = g_{x,\mathrm{lower}} + g_{x,\mathrm{upper}}$$

How the partitioning factors work:

\(k_x\)	description

0	all conductance on lower surface/portion

0.5	2/3 conductance on lower surface

1	conductance evenly divided between surfaces/portions

2	2/3 conductance on upper surface

Inf	all conductance on upper surface/portion

The boundary layer conductances for each are calculated on the basis of mass and heat transfer (see .get_gbc).

Symbol	R	Description	Units	Default

\(g_\mathrm{mc}\)	g_mc	mesophyll conductance to CO2 (T_leaf)	mol / m\(^2\) / s	calculated

\(g_\mathrm{sc}\)	g_sc	stomatal conductance to CO2	mol / m\(^2\) / s	0.4

\(g_\mathrm{uc}\)	g_uc	cuticular conductance to CO2	mol / m\(^2\) / s	0.01

\(k_\mathrm{mc}\)	k_mc	partition of \(g_\mathrm{mc}\) to lower mesophyll	none	1

\(k_\mathrm{sc}\)	k_sc	partition of \(g_\mathrm{sc}\) to lower surface	none	1

\(k_\mathrm{uc}\)	k_uc	partition of \(g_\mathrm{uc}\) to lower surface	none	1

New conductance model

The conductance model described in this section is implemented in photosynthesis (>= 2.1.0) if parameters to calculate the internal airspace and liquid-phase conductances (A_mes_A, g_liqc) are provided. These parameters are 1) the effective path lengths through the lower and upper leaf internal airspaces (delta_ias_lower, delta_ias_upper) and 2) the mesophyll area per leaf area (A_mes_A) and liquid-phase conductance per mesophyll cell area (g_liqc).

Two parallel diffusion pathways, one from each leaf surface, converge to a single CO2 concentration at the mesophyll cell boundary. We use a single liquid-phase resistance to represent the combined cell wall, plasmalemma, and chloroplast resistances. The gas-phase resistance through boundary layer, cuticle/stomata, and internal airspace is \(r_\mathrm{gas,c}\); the liquid-phase intracellular resistance is \(r_\mathrm{i,c}\).

$$r_\mathrm{total,c} = r_\mathrm{gas,c} + r_\mathrm{i,c}$$

The gas-phase resistance occurs through two parallel pathways, which we refer to as the 'lower' and 'upper' pathways because horizontally oriented leaves often have different anatomical properties on each surface. The gas-phase resistance through pathway \(i \in \{\textrm{lower,upper\}}\) is:

$$r_{\mathrm{gas,c},i} = r_{\mathrm{b,c},i} + r_{\mathrm{u+s,c},i} + r_{\mathrm{ias,c},i}$$

The subscripts \(_\mathrm{b}\), \(_\mathrm{u+s}\), and \(_\mathrm{ias}\) denote boundary layer, cuticular + stomatal, and internal airspace, respectively. The subscript \(_\mathrm{c}\) indicates we are considering the conductance to CO2 rather than another molecular species.

Cuticular and stomatal conductances (1 / resistance) are parallel, so:

$$1 / r_{\mathrm{u+s,c},i} = g_{\mathrm{u+s,c},i} = g_{\mathrm{u,c},i} + g_{\mathrm{s,c},i}$$

Substituting the above expression into the equation for \(r_{\mathrm{gas,c},i}\):

$$r_{\mathrm{gas,c},i} = r_{\mathrm{b,c},i} + 1 / (g_{\mathrm{u,c},i} = g_{\mathrm{s,c},i}) + r_{\mathrm{ias,c},i}$$

The total gas-phase resistance is the inverse of the sum of the parallel lower and upper conductances:

$$1 / r_{\mathrm{gas,c}} = g_\mathrm{gas,c,lower} + g_\mathrm{gas,c,upper}$$

The cuticular, stomatal, and mesophyll conductances can be the same or different for upper and lower. The partitioning factors \(k_u\) and \(k_s\) divide the total cuticular and stomatal conductances, respectively, between surfaces while keeping the total conductance constant:

$$g_{x,\mathrm{lower}} = g_x (1 / (1 + k_x))$$ $$g_{x,\mathrm{upper}} = g_x (k_x / (1 + k_x))$$ $$g_x = g_{x,\mathrm{lower}} + g_{x,\mathrm{upper}}$$

How the partitioning factors work:

\(k_x\)	description

0	all conductance on lower surface/portion

0.5	2/3 conductance on lower surface

1	conductance evenly divided between surfaces/portions

2	2/3 conductance on upper surface

Inf	all conductance on upper surface/portion

The internal airspace conductance is the diffusivity of CO2 at a given temperature and pressure divided by the effective path length:

$$g_\mathrm{ias,c,lower} = D_\mathrm{c} / \delta_\mathrm{ias,lower}$$ $$g_\mathrm{ias,c,upper} = D_\mathrm{c} / \delta_\mathrm{ias,upper}$$

g_iasc_lower and g_iasc_upper are calculated in the bake function. See .get_Dx for calculating D_c.

The liquid-phase intracellular resistance is given by:

$$1 / r_\mathrm{i,c} = g_\mathrm{i,c} = g_\mathrm{liq,c} A_\mathrm{mes} / A$$

\(g_\mathrm{liq,c}\) is temperature sensitive. See bake.

The boundary layer conductances for each are calculated on the basis of mass and heat transfer (see .get_gbc).