Convert g_c (\(\mu\)mol CO2/m^2/s/Pa) to g_w (\(\mu\)mol H2O /m^2/s/Pa)

Convert g_c (\(\mu\)mol CO2/m^2/s/Pa) to g_w (\(\mu\)mol H2O /m^2/s/Pa)

Convert g_c (umol CO2/m^2/s/Pa) to g_w (umol H2O /m^2/s/Pa)

gw2gc(g_w, D_c, D_w, unitless)

gc2gw(g_c, D_c, D_w, unitless)

Arguments

g_w
conductance to water vapor in units (\(\mu\)mol H2O / (m^2 s Pa)) of class units.

D_c
diffusion coefficient for CO2 in air in units of m^2/s of call units

D_w
diffusion coefficient for H2O in air in units of m^2/s of call units

unitless
Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

g_c
conductance to CO2 in units (\(\mu\)mol H2O / (m^2 s Pa)) of class units.

Value

Value with units \(\mu\)mol / (m^2 s Pa) of class units.

Details

Diffusive conductance to CO2 is generally about ~1.6x that of H2O because of the higher molecular weight. To convert, multiply conductance by the ratio of diffusion coefficients:

$$g_\mathrm{c} = g_\mathrm{w} D_\mathrm{c} / D_\mathrm{w}$$ $$g_\mathrm{w} = g_\mathrm{c} D_\mathrm{w} / D_\mathrm{c}$$

Note

This function will soon be moving to the standalone gunit package.

Examples

D_c <- set_units(1.29e-05, "m^2/s")
D_w <- set_units(2.12e-05, "m^2/s")
g_c <- set_units(3, "umol/m^2/s/Pa")
g_w <- gc2gw(g_c, D_c, D_w, unitless = FALSE)
g_w

#> 4.930233 [umol/m^2/Pa/s]

gw2gc(g_w, D_c, D_w, unitles = FALSE)

#> 3 [umol/m^2/Pa/s]