Make lists of parameters for photosynthesis

Make lists of parameters for photosynthesis

make_leafpar

make_enviropar

make_bakepar

make_constants

make_leafpar(replace = NULL)

make_enviropar(replace = NULL)

make_bakepar(replace = NULL)

make_constants(replace = NULL)

Arguments

replace
A named list of parameters to replace defaults. If NULL, defaults will be used.

Value

make_leafpar: An object inheriting from class leaf_par
make_enviropar: An object inheriting from class enviro_par
make_bakepar: An object inheriting from class bake_par
make_constants: An object inheriting from class constants

Details

Leaf parameters:

Symbol			R

Description

Units			Default

\(d\)			leafsize

leaf characteristic dimension	m	0.1

\(\Gamma*\)			gamma_star

chloroplastic CO2 compensation point (T_leaf)	Pa		calculated

\(\Gamma*_{25}\)			gamma_star25

chloroplastic CO2 compensation point (25 °C)	Pa	3.743

\(g_\mathrm{mc}\)			g_mc

mesophyll conductance to CO2 (T_leaf)		\(\mu\)	mol CO2 / (m	\(^2\)

s Pa)		calculated

\(g_\mathrm{mc}\)			g_mc25

mesophyll conductance to CO2 (25 °C)		\(\mu\)	mol CO2 / (m	\(^2\)

s Pa)	4		\(g_\mathrm{sc}\)

g_sc		stomatal conductance to CO2

\(\mu\)	mol CO2 / (m	\(^2\)	s Pa)	4

\(g_\mathrm{uc}\)			g_uc

cuticular conductance to CO2		\(\mu\)	mol CO2 / (m

\(^2\)	s Pa)	0.1		\(J_\mathrm{max25}\)

J_max25		potential electron transport (25 °C)

\(\mu\)	mol CO2 / (m	\(^2\)	s)

200		\(J_\mathrm{max}\)

J_max		potential electron transport (T_leaf)		\(\mu\)

mol CO2 / (m	\(^2\)	s)		calculated

\(k_\mathrm{mc}\)

k_mc		partition of	\(g_\mathrm{mc}\)	to lower mesophyll

none	1		\(k_\mathrm{sc}\)

k_sc		partition of	\(g_\mathrm{sc}\)

to lower surface	none	1		\(k_\mathrm{uc}\)

k_uc		partition of

\(g_\mathrm{uc}\)	to lower surface	none	1

\(K_\mathrm{C25}\)			K_C25

Michaelis constant for carboxylation (25 °C)		\(\mu\)	mol / mol	268.3

\(K_\mathrm{C}\)			K_C

Michaelis constant for carboxylation (T_leaf)		\(\mu\)	mol / mol

calculated			\(K_\mathrm{O25}\)

K_O25		Michaelis constant for oxygenation (25 °C)

\(\mu\)	mol / mol	165084.2

\(K_\mathrm{O}\)			K_O

Michaelis constant for oxygenation (T_leaf)		\(\mu\)	mol / mol

calculated			\(\phi_J\)

phi_J		initial slope of the response of J to PPFD	none

0.331		\(R_\mathrm{d25}\)

R_d25		nonphotorespiratory CO2 release (25 °C)		\(\mu\)

mol CO2 / (m	\(^2\)	s)	2

\(R_\mathrm{d}\)			R_d

nonphotorespiratory CO2 release (T_leaf)		\(\mu\)	mol CO2 / (m	\(^2\)

s)		calculated

\(\theta_J\)			theta_J

curvature factor for light-response curve	none	0.825		\(T_\mathrm{leaf}\)

T_leaf		leaf temperature

K	298.15		\(V_\mathrm{c,max25}\)

V_cmax25		maximum rate of carboxylation (25 °C)

\(\mu\)	mol CO2 / (m	\(^2\)	s)	150

\(V_\mathrm{c,max}\)			V_cmax

maximum rate of carboxylation (T_leaf)		\(\mu\)	mol CO2 / (m

\(^2\)	s)		calculated

\(V_\mathrm{tpu25}\)			V_tpu25

rate of triose phosphate utilisation (25 °C)		\(\mu\)	mol CO2 / (m

\(^2\)	s)	200		\(V_\mathrm{tpu}\)

V_tpu		rate of triose phosphate utilisation (T_leaf)

\(\mu\)	mol CO2 / (m	\(^2\)	s)

calculated			Symbol

Environment parameters:

Symbol			R

Description

Units			Default

\(C_\mathrm{air}\)			C_air

atmospheric CO2 concentration	Pa	41

\(O\)			O

atmospheric O2 concentration	kPa	21.27565		\(P\)

P		atmospheric pressure

kPa	101.3246		PPFD

PPFD		photosynthetic photon flux density	umol quanta / (m^2 s)	1500

\(\mathrm{RH}\)			RH

relative humidity	none	0.50

\(T_\mathrm{air}\)			T_air

air temperature	K	298.15		\(u\)

wind		windspeed

m / s	2		Symbol

Baking (i.e. temperature response) parameters:

Symbol			R

Description

Units			Default

\(D_\mathrm{s,gmc}\)			Ds_gmc

empirical temperature response parameter	J / (mol K)	487.29

\(D_\mathrm{s,Jmax}\)			Ds_Jmax

empirical temperature response parameter	J / (mol K)	388.04		\(E_\mathrm{a,\Gamma *}\)

Ea_gammastar		empirical temperature response parameter

J / mol	24459.97		\(E_\mathrm{a,gmc}\)

Ea_gmc		empirical temperature response parameter	J / mol

68901.56		\(E_\mathrm{a,Jmax}\)

Ea_Jmax		empirical temperature response parameter	J / mol	56095.18

\(E_\mathrm{a,KC}\)			Ea_KC

empirical temperature response parameter	J / mol	80989.78

\(E_\mathrm{a,KO}\)			Ea_KO

empirical temperature response parameter	J / mol	23719.97		\(E_\mathrm{a,Rd}\)

Ea_Rd		empirical temperature response parameter

J / mol	40446.75		\(E_\mathrm{a,Vcmax}\)

Ea_Vcmax		empirical temperature response parameter	J / mol

52245.78		\(E_\mathrm{d,gmc}\)

Ed_gmc		empirical temperature response parameter	J / mol	148788.56

\(E_\mathrm{d,Jmax}\)			Ed_Jmax

empirical temperature response parameter	J / mol	121244.79

Constants:

Symbol			R

Description

Units			Default

\(D_{c,0}\)			D_c0

diffusion coefficient for CO2 in air at 0 °C	m	\(^2\)	/ s

12.9		\(D_{h,0}\)

D_h0		diffusion coefficient for heat in air at 0 °C	m	\(^2\)

/ s	1.9e-5		\(D_{m,0}\)

D_m0		diffusion coefficient for momentum in air at 0 °C	m

\(^2\)	/ s	13.3e-06		\(\epsilon\)

epsilon		ratio of water to air molar masses

none	0.622		\(G\)

G		gravitational acceleration	m / s

\(^2\)		9.8		\(eT\)

eT		exponent for temperature dependence of diffusion

none	1.75		\(R\)

R		ideal gas constant	J / (mol K)

8.3144598		\(\sigma\)

s		Stephan-Boltzmann constant	W / (m	\(^2\)

K	\(^4\)	)	5.67e-08

\(Sh\)			Sh

Sherwood number	none		calculated

References

Buckley TN and Diaz-Espejo A. 2015. Partitioning changes in photosynthetic rate into contributions from different variables. Plant, Cell & Environment 38: 1200-11.

Examples

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()

leaf_par <- make_leafpar(
  replace = list(
    g_sc = set_units(3, "umol/m^2/s/Pa"),
    V_cmax25 = set_units(100, "umol/m^2/s")
  )
)