Simulate C3 photosynthesis

photosynthesis: simulate C3 photosynthesis over multiple parameter sets

photo: simulate C3 photosynthesis over a single parameter set

photosynthesis(leaf_par, enviro_par, bake_par, constants,
  progress = TRUE, quiet = FALSE, set_units = TRUE,
  parallel = FALSE)

photo(leaf_par, enviro_par, bake_par, constants, quiet = FALSE,
  set_units = TRUE)

Arguments

leaf_par
A list of leaf parameters inheriting class leaf_par. This can be generated using the make_leafpar function.

enviro_par
A list of environmental parameters inheriting class enviro_par. This can be generated using the make_enviropar function.

bake_par
A list of temperature response parameters inheriting class bake_par. This can be generated using the make_bakepar function.

constants
A list of physical constants inheriting class constants. This can be generated using the make_constants function.

progress
Logical. Should a progress bar be displayed?

quiet
Logical. Should messages be displayed?

set_units
Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.

parallel
Logical. Should parallel processing be used via future_map?

Value

A data.frame with the following units columns

Input:

C_air

atmospheric CO2 concentration (Pa)

g_mc25

mesophyll conductance to CO2 at 25 °C (

\(\mu\)	mol CO2 / (m

\(^2\)	s Pa))

g_sc

stomatal conductance to CO2 (

\(\mu\)	mol CO2 / (m

\(^2\)	s Pa))

g_uc

cuticular conductance to CO2 (

\(\mu\)	mol CO2 / (m

\(^2\)	s Pa))

gamma_star25

chloroplastic CO2 compensation point at 25 °C (Pa)

J_max25

potential electron transport at 25 °C (

\(\mu\)	mol CO2) / (m

\(^2\)	s)

K_C25

Michaelis constant for carboxylation at 25 °C (

\(\mu\)	mol / mol)

K_O25

Michaelis constant for oxygenation at 25 °C (

\(\mu\)	mol / mol)

k_mc

partition of

\(g_\mathrm{mc}\)	to lower mesophyll (unitless)

k_sc

partition of

\(g_\mathrm{sc}\)	to lower surface (unitless)

k_uc

partition of

\(g_\mathrm{uc}\)	to lower surface (unitless)

leafsize

leaf characteristic dimension (m)

O

atmospheric O2 concentration (kPa)

P

atmospheric pressure (kPa)

phi_J

initial slope of the response of J to PPFD (unitless)

PPFD

photosynthetic photon flux density (umol quanta / (m

\(^2\)	s))

R_d25

nonphotorespiratory CO2 release at 25 °C (

\(\mu\)	mol CO2 / (m

\(^2\)	s))

RH

relative humidity (unitless)

theta_J

curvature factor for light-response curve (unitless)

T_air

air temperature (K)

T_leaf

leaf tempearture (K)

V_cmax25

maximum rate of carboxylation at 25 °C (

\(\mu\)	mol CO2 / (m

\(^2\)	s))

V_tpu25

rate of triose phosphate utilisation at 25 °C (

\(\mu\)	mol CO2 / (m

\(^2\)	s))

wind

wind speed (m / s)

Baked Input:

g_mc

mesophyll conductance to CO2 at

T_leaf	(

\(\mu\)	mol CO2 / (m

\(^2\)	s Pa))

gamma_star

chloroplastic CO2 compensation point at

T_leaf	(Pa)

J_max

potential electron transport at

T_leaf	(

\(\mu\)	mol CO2) / (m

\(^2\)	s)

K_C

Michaelis constant for carboxylation at

T_leaf	(

\(\mu\)	mol / mol)

K_O

Michaelis constant for oxygenation at

T_leaf	(

\(\mu\)	mol / mol)

R_d

nonphotorespiratory CO2 release at

T_leaf	(

\(\mu\)	mol CO2 / (m

\(^2\)	s))

V_cmax

maximum rate of carboxylation at

T_leaf	(

\(\mu\)	mol CO2 / (m

\(^2\)	s))

V_tpu

rate of triose phosphate utilisation at

T_leaf	(

\(\mu\)	mol CO2 / (m

\(^2\)	s))

Output:

A

photosynthetic rate at

C_chl	(

\(\mu\)	mol CO2 / (m

\(^2\)	s))

C_chl

chloroplastic CO2 concentration where

A_supply	intersects

A_demand	(Pa)

g_tc

total conductance to CO2 at

T_leaf	(

\(\mu\)	mol CO2 / (m

\(^2\)	s Pa))

value

A_supply	-

A_demand	(

\(\mu\)	mol CO2 / (m

\(^2\)	s)) at

C_chl

convergence

convergence code (0 = converged)

Details

photo: This function takes simulates photosynthetic rate using the Farquhar-von Caemmerer-Berry (FvCB) model of C3 photosynthesis for single combined set of leaf parameters (leaf_par), environmental parameters (enviro_par), and physical constants (constants). Leaf parameters are provided at reference temperature (25 °C) and then "baked" to the appropriate leaf temperature using temperature response functions (see bake).

photosynthesis: This function uses photo to simulate photosynthesis over multiple parameter sets that are generated using crossing.

Examples

# Single parameter set with 'photo'

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()
photo(leaf_par, enviro_par, bake_par, constants)

#> 
#> Solving for C_chl ...

#>  done

#>          C_chl         value convergence                     g_tc
#> 1 24.9227 [Pa] -1.161979e-06           0 1.721631 [umol/m^2/Pa/s]
#>                       A            g_mc25              g_sc                g_uc
#> 1 27.67919 [umol/m^2/s] 4 [umol/m^2/Pa/s] 4 [umol/m^2/Pa/s] 0.1 [umol/m^2/Pa/s]
#>   gamma_star25          J_max25       K_C25        K_O25  k_mc  k_sc  k_uc
#> 1   3.743 [Pa] 200 [umol/m^2/s] 27.238 [Pa] 16.582 [kPa] 1 [1] 1 [1] 1 [1]
#>   leafsize     phi_J          R_d25     T_leaf   theta_J         V_cmax25
#> 1  0.1 [m] 0.331 [1] 2 [umol/m^2/s] 298.15 [K] 0.825 [1] 150 [umol/m^2/s]
#>            V_tpu25 g_mc gamma_star J_max    K_C    K_O R_d V_cmax V_tpu   C_air
#> 1 200 [umol/m^2/s]    4      3.743   200 27.238 16.582   2    150   200 41 [Pa]
#>                O              P              PPFD      RH      T_air    wind
#> 1 21.27565 [kPa] 101.3246 [kPa] 1500 [umol/m^2/s] 0.5 [1] 298.15 [K] 2 [m/s]

# Multiple parameter sets with 'photosynthesis'

leaf_par <- make_leafpar(
  replace = list(
    T_leaf = set_units(c(293.14, 298.15), "K")
    )
  )
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()
photosynthesis(leaf_par, enviro_par, bake_par, constants)

#> Solving for photosynthetic rate from 2 parameter sets ...

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

#>      C_chl         value convergence     g_tc        A            g_mc25
#> 1 26.05248 -3.271743e-06           0 1.493755 22.32793 4 [umol/m^2/Pa/s]
#> 2 24.92270 -1.161979e-06           0 1.721631 27.67919 4 [umol/m^2/Pa/s]
#>                g_sc                g_uc gamma_star25          J_max25
#> 1 4 [umol/m^2/Pa/s] 0.1 [umol/m^2/Pa/s]   3.743 [Pa] 200 [umol/m^2/s]
#> 2 4 [umol/m^2/Pa/s] 0.1 [umol/m^2/Pa/s]   3.743 [Pa] 200 [umol/m^2/s]
#>         K_C25        K_O25  k_mc  k_sc  k_uc leafsize     phi_J          R_d25
#> 1 27.238 [Pa] 16.582 [kPa] 1 [1] 1 [1] 1 [1]  0.1 [m] 0.331 [1] 2 [umol/m^2/s]
#> 2 27.238 [Pa] 16.582 [kPa] 1 [1] 1 [1] 1 [1]  0.1 [m] 0.331 [1] 2 [umol/m^2/s]
#>       T_leaf   theta_J         V_cmax25          V_tpu25   C_air              O
#> 1 293.14 [K] 0.825 [1] 150 [umol/m^2/s] 200 [umol/m^2/s] 41 [Pa] 21.27565 [kPa]
#> 2 298.15 [K] 0.825 [1] 150 [umol/m^2/s] 200 [umol/m^2/s] 41 [Pa] 21.27565 [kPa]
#>                P              PPFD      RH      T_air    wind
#> 1 101.3246 [kPa] 1500 [umol/m^2/s] 0.5 [1] 298.15 [K] 2 [m/s]
#> 2 101.3246 [kPa] 1500 [umol/m^2/s] 0.5 [1] 298.15 [K] 2 [m/s]
#>                       g_mc    gamma_star                 J_max           K_C
#> 1 2.844658 [umol/m^2/Pa/s] 3.162148 [Pa] 143.6921 [umol/m^2/s] 15.58393 [Pa]
#> 2 4.000000 [umol/m^2/Pa/s] 3.743000 [Pa] 200.0000 [umol/m^2/s] 27.23800 [Pa]
#>             K_O                 R_d                V_cmax                 V_tpu
#> 1 14.0804 [kPa] 1.5133 [umol/m^2/s] 104.6304 [umol/m^2/s] 139.5072 [umol/m^2/s]
#> 2 16.5820 [kPa] 2.0000 [umol/m^2/s] 150.0000 [umol/m^2/s] 200.0000 [umol/m^2/s]
#>   g_mc251 g_sc1 g_uc1 gamma_star251 J_max251 K_C251 K_O251 k_mc1 k_sc1 k_uc1
#> 1       4     4   0.1         3.743      200 27.238 16.582     1     1     1
#> 2       4     4   0.1         3.743      200 27.238 16.582     1     1     1
#>   leafsize1 phi_J1 R_d251 T_leaf1 theta_J1 V_cmax251 V_tpu251 C_air1       O1
#> 1       0.1  0.331      2  293.14    0.825       150      200     41 21.27565
#> 2       0.1  0.331      2  298.15    0.825       150      200     41 21.27565
#>         P1 PPFD1 RH1 T_air1 wind1
#> 1 101.3246  1500 0.5 298.15     2
#> 2 101.3246  1500 0.5 298.15     2